seq: use guint8 for the numerical ID of port

+++ src/seq/query.h
@@ -27,10 +27,10 @@ void alsaseq_get_client_id_list(guint **entries, gsize *entry_count,
 void alsaseq_get_client_info(guint8 client_id, ALSASeqClientInfo **client_info,
 			     GError **error);
 
-void alsaseq_get_port_id_list(guint8 client_id, guint **entries,
+void alsaseq_get_port_id_list(guint8 client_id, guint8 **entries,
                               gsize *entry_count, GError **error);
 
-void alsaseq_get_port_info(guint8 client_id, guint port_id,
+void alsaseq_get_port_info(guint8 client_id, guint8 port_id,
                            ALSASeqPortInfo **port_info, GError **error);
 
 void alsaseq_get_client_pool(guint8 client_id, ALSASeqClientPool **client_pool,
